requirements.txt
absl_py==2.1.0
accelerate==0.32.1
braceexpand==0.1.7
# Calvin.egg==info
# calvin_env==0.0.1
git+https://github.com/moojink/dlimp_openvla
draccus==0.8.0
einops==0.8.1
ema_pytorch==0.5.1
gym==0.26.2
h5py==3.11.0
huggingface_hub==0.26.1
hydra-core==1.3.2
imageio==2.34.2
jsonlines==4.0.0
lightning==2.4.0
matplotlib==3.10.1
moviepy==1.0.3
numpy==1.26.4
omegaconf==2.3.0
opencv_python==4.10.0.84
# openvla==0.0.3
packaging==24.1
peft==0.11.1
Pillow==11.2.1
piq==0.8.0
pyquaternion==0.9.9
pytorch_lightning==1.8.6
PyYAML==6.0.1
Requests==2.32.3
rich==14.0.0
robosuite==1.4.1
rotary_embedding_torch==0.8.4
setuptools==57.5.0
tensorflow==2.15.0
tensorflow_datasets==4.9.3
tensorflow_graphics==2021.12.3
termcolor==3.0.1
timm==0.9.10
tokenizers==0.19.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.4
transformers==4.40.1
webdataset==0.2.111
